Fix server env accessed from client
