+++ src/server/external.ts
@@ -3,7 +3,6 @@ import '@kiltprotocol/augment-api'
 import { typesBundle } from '@kiltprotocol/type-definitions'
 import { ApiPromise, HttpProvider } from '@polkadot/api'
 import axios from 'axios'
-import { GraphQLClient, RequestOptions, Variables } from 'graphql-request'
 
 export const covalentRequest = axios.create({
   baseURL: 'https://api.covalenthq.com/v1/',
@@ -22,22 +21,4 @@ export const getKiltApi = async () => {
   kiltApi = api
 
   return api
-}
-
-export function getSubIdRequest() {
-  return axios.create({ baseURL: 'https://sub.id/api/v1/' })
-}
-
-export function subsocialSquidRequest<T, V extends Variables = Variables>(
-  config: RequestOptions<V, T>
-) {
-  const url = 'https://squid.subsquid.io/subsocial/graphql'
-
-  const SQUID_TIMEOUT = 3 * 1000 // 3 seconds
-  const client = new GraphQLClient(url, {
-    timeout: SQUID_TIMEOUT,
-    ...config,
-  })
-
-  return client.request({ url, ...config })
 }
